requirements.txt
pytest
allure_pytest
requests
dateutils
pandas
pytest-xdist
pytest-rerunfailures
pytest-selenium